[FE] Refactoring : 에러핸들링 수정 (#247)
* .gitignore .p12파일 추가

* subCategory 눌렀다가 mainCategory 눌렀을 때 제대로 적용되도록 (#198)


* feat: 카테고리 관련 구직데이터 크롤링 v1 (#199)

* fix: 데이터 갱신시 이전 데이터 삭제처리 (#201)

* [FE] Feat : answer comment 작성 기능 구현, PostPage CRUD시에 렌더링 되도록 구현, 헤더 마이페이지 접속 에러 해결, 헤더 마이페이지 접속시 drop-down 닫히도록 구현. (#200)

* feat: 유저 점수 추가 (#207)

* 이메일 발송 기능 구현 (#208)

* Feat: 이메일 발송 기본 코드 작성 및 테스트

* Feat: 회원가입 메일 인증, 비밀번호 분실 시 임시 비밀번호 발급 기능 구현

* feat: Recruit테이블에 Auditable 확장

* fix: 유저 삭제시 뱃지 삭제

* fix: 로그인 및 인증 ErrorResponse 추가 (#210)

* fix: 로그인 및 인증 ErrorResponse 추가

* fix: ErrorResponse 일부 한글메시지 변경 및 스케줄 메시지 추가

* fix: ErrorResponse

* fix: 스케줄 시간 수정

* fix: 로그인 에러처리 추가

* fix: 에러 코드 수정

* comment 수정, 삭제 기능 구현 (#209)

* [FE] Feat : answer comment 작성 기능 구현, PostPage CRUD시에 렌더링 되도록 구현, 헤더 마이페이지 접속 에러 해결, 헤더 마이페이지 접속시 drop-down 닫히도록 구현.

* [FE] Feat : 댓글 수정, 삭제 기능 구현 완료.

* [Fe] Feat : 내 질문이 아닌 다른 질문에서 내가 작성한 comment 안보이는 로직 수정.

* 마이페이지(유저페이지) 정보변경기능 추가 (#211)

* subCategory 눌렀다가 mainCategory 눌렀을 때 제대로 적용되도록

* 프로필사진 변경 가능

* err message data 그대로 나오도록 변경

* 댓글순 삭제

* Python,JPA 백엔드 추가

* 비밀번호 바꾸기

* 회원정보창은 프로필이미지 수정 불가능하고 내 계정은 수정 가능하도록

* 내 계정 탭은 내 id를 조회할 때만 볼 수 있도록

* 이미지변경 text 위치 조정

* Refactor : UserTab id 사용하고 있던 요소들을 고침

* 캐러셀 삭제

* 글자를 화이트로 변경

* [FE] Fix : PostPage에서 글자수가 많아지면 겹치는 현상 수정 (#212)

* 이미지 크롭 기능 추가 (#214)

* fix: 좋아요 권한처리 수정 (#216)

* Update README.md

* feat: 질문 검색 기능에 정렬 추가 (#217)

* Feat/search (#220)



* [FE] Refac : 답변,댓글 수정 후 리렌더링 되도록 수정

* [FE] Feat :  검색기능 구현 중, subcategory 오류

* bug 수정 (#221)


* Refactoring 예시 (#223)


* feat: 대용양 데이터 조회 속도 개선 (#222)

* Refactor/directory (#224)

* Refactoring 예시

* Comment,User axiosRequest 리팩토링

* axiosRequest 마무리

* eslint로 unused-import 삭제 설정


* fix: 공통세팅 분류 (#225)

* fix: 공통세팅 분류

* fix: 공통세팅 분류 수정

* Refactor/error handling (#226)


* 만약 로그인실패시 SignInModal이 안닫히도록

* 에러핸들링 커버

* 질문 작성 시 질문이 있는지 검색해보라는 문구 작성


* fix: 불필요한 테스트 삭제

* [FE] Refactor : 로그인 모달 리팩토링 (#228)

* SignInModal 어디서든 열릴수 있도록 refactoring

* 로그인을 했을 때 메인페이지로 바로 가는것보단 그냥 현재 페이지에 있는것이 좋아보임

* 로그인 안한 상태에서 질문/답변/댓글 작성시 로그인모달 띄우기


* [FE] Refactor : QuestionCard 사이 마진 클릭 안되도록 수정 (#229)

* SignInModal 어디서든 열릴수 있도록 refactoring

* 로그인을 했을 때 메인페이지로 바로 가는것보단 그냥 현재 페이지에 있는것이 좋아보임

* 로그인 안한 상태에서 질문/답변/댓글 작성시 로그인모달 띄우기

* QuestionCard margin 사이 클릭 안되도록 변경


* [FE] Style : QuestionCard UI 개선 (#230)

* SignInModal 어디서든 열릴수 있도록 refactoring

* 로그인을 했을 때 메인페이지로 바로 가는것보단 그냥 현재 페이지에 있는것이 좋아보임

* 로그인 안한 상태에서 질문/답변/댓글 작성시 로그인모달 띄우기

* QuestionCard margin 사이 클릭 안되도록 변경

* QuestionCard UI 개선


* [FE] Style : 브라우저에서 볼 때 이름 변경 (#231)

* SignInModal 어디서든 열릴수 있도록 refactoring

* 로그인을 했을 때 메인페이지로 바로 가는것보단 그냥 현재 페이지에 있는것이 좋아보임

* 로그인 안한 상태에서 질문/답변/댓글 작성시 로그인모달 띄우기

* QuestionCard margin 사이 클릭 안되도록 변경

* QuestionCard UI 개선

* TabName 수정(원래 react app이라고 떴음)


* 질문 등록 후 폼 리셋. (#234)

* [FE]  Refactor : 질문 등록 후 form 리셋.

* no message

* refactor: 필드에러와 일반 에러 형식 통일 (#237)

* 로그아웃시 alert 띄우는것 구현. (#238)


* [FE] Refactor : 로그아웃시 유저에게 alert 띄우기 구현, 로그인 없이 답변작성과 질문등록시 로직 수정.

* [FE] Refactor : 로직수정.

* 페이지 이동시 최상단으로 이동하도록 구현. (#239)

* [FE] Refactor : 로그아웃시 유저에게 alert 띄우기 구현, 로그인 없이 답변작성과 질문등록시 로직 수정.

* [FE] Refactor : 로직수정.

* [FE] Feat : 페이지 이동시 최상단으로 가도록 구현. 답변등록, 질문등록 로직 수정.

* [FE] Refactor : 불필요한 로직 삭제.

* 회원탈퇴기능 추가 (#240)


* [FE] Refactor : UI (#241)

* 회원탈퇴기능 추가

* scrollToTop 버튼 생성

* subCategory 이미지 삭제

* questionCard UI 살짝 수정


* 코멘트에 나오는 프로필 수정, 회원탈퇴시 alert주고 로그아웃까지

* 캐러셀 완성

* 임의로 transform 조작시 clearInterval

* useInterval 이라는 커스텀훅 만들어서 사용

* useInterval(자동재생) 제거

* postSignIn error 핸들링 개편 및 로그인실패시 password reset

* postSignUp error 핸들링 개편

* question,answer,comment post 및 updateAnswerVotes error 핸들링 개편
