+++ client/src/api/Answer/delete.js
@@ -4,6 +4,7 @@ export async function deleteAnswer(answerId) {
     const res = await axiosInstance.delete(`/answers/${answerId}`);
     alert('삭제되었습니다');
   } catch (err) {
-    return err;
+    alert(err.response.data.message);
+    throw new Error(err.response.data.status);
   }
 }
