+++ client/src/api/Comment/post.js
@@ -6,6 +6,7 @@ export async function postComment(answerId, data) {
     alert('댓글 작성이 완료되었습니다');
     return res;
   } catch (err) {
-    return err;
+    alert(err.response.data.message);
+    throw new Error(err.response.data.status);
   }
 }
