+++ client/src/api/Question/post.js
@@ -3,8 +3,8 @@ export async function postQuestion(data) {
   try {
     const res = await axiosInstance.post('/questions', data);
     alert('질문 작성이 완료되었습니다');
-    return res;
   } catch (err) {
-    return err;
+    alert(err.response.data.message);
+    throw new Error(err.response.data.status);
   }
 }
