+++ client/src/api/User/signUp.js
@@ -3,8 +3,8 @@ import axiosInstance from '../../utils/axiosInstance';
 export async function postSignUp(data) {
   try {
     const res = await axiosInstance.post('/users', data);
-    return res;
   } catch (err) {
-    throw new Error('err');
+    alert(err.response.data.message);
+    throw new Error(err.response.data.status);
   }
 }
