+++ client/src/utils/axiosInstance.js
@@ -26,15 +26,6 @@ axiosInstance.interceptors.response.use(
   },
   async (error) => {
     if (
-      error?.response?.data?.status === 404 &&
-      error?.response?.data?.message === '탈퇴한 유저입니다.'
-    ) {
-      alert('탈퇴한 유저입니다');
-      postLogout();
-    }
-    if (error?.response?.status === 403) {
-      alert('로그인 해주세요.');
-    } else if (
       error?.response?.status === 401 &&
       error?.response?.data?.message === 'ACCESS TOKEN EXPIRED'
     ) {
